Back End/requirements.txt
django
django-cors-headers
django-import-export
cycler
matplotlib
numpy
pyparsing
python-dateutil
pytz
scipy
scikit-learn
six
xlrd
tensorflow
